fix: useSummary hook
