+++ src/components/summary/index.tsx
@@ -4,31 +4,11 @@ import {
   CurrencyDollar
 } from '@phosphor-icons/react'
 import { SummaryCard, SummaryContainer } from './styles'
-import { useContext } from 'react'
-import { TransactionsContext } from '../../contexts/transactionscontext'
 import { priceFormatter } from '../../utils/formatter'
+import { useSummary } from '../../hooks/useSummary'
 
 export function Summary() {
-  const { transactions } = useContext(TransactionsContext)
-
-  const summary = transactions.reduce(
-    (acc, transaction) => {
-      if (transaction.type === 'income') {
-        acc.income += transaction.price
-        acc.total += transaction.price
-      } else {
-        acc.outcome += transaction.price
-        acc.total -= transaction.price
-      }
-
-      return acc
-    },
-    {
-      income: 0,
-      outcome: 0,
-      total: 0
-    }
-  )
+  const summary = useSummary()
 
   return (
     <SummaryContainer>
